fix: use declare keyword in the .d.ts files (#10)
The dts-bundle project is not maintained anymore. There is a PR to fix an issue but no-one merges it. I could merge it as I'm part of the TypeStrong organization but I don't have a publish access.  To overcome this blocker, I've copied the package and applied the fix. I didn't found any maintained package that could inline typings as dts-bundle does.

✅ Closes: #8